4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -35,6 +35,22 @@ blogRouter.post("/", async (request, response) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const blogFound = await Blog.findById(request.params.id);
+
+  if (!blogFound) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+
+  if (blogFound.user.toString() !== decodedToken.id) {
+    return response.status(403).json({ error: "user cannot delete this blog" });
+  }
+
   const result = await Blog.findByIdAndRemove(request.params.id);
 
   if (result) {
